Documentation improvements. Main script now prompts for all needed parameters.

+++ nuodbcluster/cluster.py
@@ -6,25 +6,41 @@ Created on Jan 28, 2014
 
 import boto.route53
 import nuodbaws
-import inspect, json, os, random, shelve, string, sys, time
+import inspect, json, os, random, string, sys, time
 
 class NuoDBCluster:
     
-    def __init__(self, aws_access_key = "", aws_secret = "", cluster_name = "default", 
-                 dns_domain="", domain_name="domain", domain_password="bird", 
-                 instance_type = "m1.large", nuodb_license = "", ssh_key = "",  
+    def __init__(self, 
+                 alert_email = "alert@example.com",
+                 aws_access_key = "", 
+                 aws_secret = "", 
+                 brokers_per_zone = 2,
+                 cluster_name = "default",
                  data_dir = "/".join([os.path.dirname(os.path.abspath(inspect.stack()[-1][1])), "data"]), 
-                 brokers_per_zone = 2, enable_monitoring = True, alert_email = "alert@example.com"):
+                 dns_domain="", 
+                 domain_name="domain", 
+                 domain_password="bird", 
+                 enable_monitoring = True,
+                 instance_type = "m1.large", 
+                 nuodb_license = "", 
+                 ssh_key = "", 
+                 ssh_keyfile = None):
       self.route53 = boto.route53.connection.Route53Connection(aws_access_key_id=aws_access_key, aws_secret_access_key=aws_secret)
-      database_file = "/".join([data_dir, cluster_name + ".shelf"])
       args, _, _, values = inspect.getargvalues(inspect.currentframe())
       for i in args:
         setattr(self, i, values[i])
-      
-      self.database_file = database_file
-      self.db = shelve.open(database_file, writeback = True)
+        
+      if ssh_keyfile != None and ssh_keyfile != "":
+        if not os.path.exists(ssh_keyfile):
+          raise Error("Can not find ssh private key %s" % self.ssh_keyfile)
+      if dns_domain == None or dns_domain == "None":
+        self.dns_domain = "nuoDB"
+        self.dns_emulate = True
+      else:
+        self.dns_emulate = False
+        
+      self.db = {}
       self.zones = {} #store our zone connections
-      #self.db.close()
     
     def add_host(self, name, zone, ami = "", security_group_ids=[], subnets = [], agentPort = 48004 , subPortRange = 48005, nuodb_rpm_url = None):
       if zone not in self.zones:
@@ -50,16 +66,20 @@ class NuoDBCluster:
         stub[host] = {}
       # Generate data for chef... is it a broker? peers?
       agent_addr = fqdn
+      if "zones" not in self.db['customers'][self.cluster_name]:
+        self.db['customers'][self.cluster_name]['zones'] = {}
+      if zone not in self.db['customers'][self.cluster_name]['zones']:
+        self.db['customers'][self.cluster_name]['zones'][zone] = {"brokers": []}
       if "chef_data" not in stub[host]:
-        if "brokers" not in self.db['customers'][self.cluster_name] or len(self.db['customers'][self.cluster_name]['brokers']) < 1:
+        if len(self.db['customers'][self.cluster_name]['zones'][zone]['brokers']) < 1:
           isBroker = True
-          chef_data = {"nuodb": {"is_broker": True, "enableAutomation": True, "enableAutomationBootstrap": True}}
-          self.db['customers'][self.cluster_name]['brokers'] = [agent_addr]
+          chef_data = {"nuodb": {"is_broker": True, "enableAutomation": True, "enableAutomationBootstrap": True, "autoconsole": {"brokers": ["localhost"]}}}
+          #self.db['customers'][self.cluster_name]['brokers'] = [agent_addr]
           self.db['customers'][self.cluster_name]['zones'][zone]['brokers'] =[agent_addr]
-        elif len(self.db['customers'][self.cluster_name]['zones'][zone]['brokers']) < self.brokers_per_zone:
+        elif len(self.db['customers'][self.cluster_name]['zones'][zone]['brokers']) < int(self.brokers_per_zone):
           isBroker = True
-          chef_data = {"nuodb": {"is_broker": True, "enableAutomation": False, "enableAutomationBootstrap": False}}
-          self.db['customers'][self.cluster_name]['brokers'].append(agent_addr)
+          chef_data = {"nuodb": {"is_broker": True, "enableAutomation": False, "enableAutomationBootstrap": False, "autoconsole": {"brokers": ["localhost"]}}}
+          #self.db['customers'][self.cluster_name]['brokers'].append(agent_addr)
           self.db['customers'][self.cluster_name]['zones'][zone]['brokers'].append(agent_addr)
         else:
           isBroker = False
@@ -89,10 +109,12 @@ class NuoDBCluster:
                                              domain = self.domain_name, domainPassword = self.domain_password, 
                                              advertiseAlt = True, region = zone,
                                              agentPort = agentPort, portRange = subPortRange,
-                                             isBroker = isBroker)
+                                             isBroker = isBroker, ssh_key = self.ssh_key, ssh_keyfile = self.ssh_keyfile)
       return host
 
-    def __boot_host(self, host, zone):
+    def __boot_host(self, host, zone, instance_type = None, wait_for_health = False):
+      if instance_type == None:
+        instance_type = self.instance_type
       stub = self.db['customers'][self.cluster_name]['zones'][zone]['hosts'][host]
       template_vars = dict(
                           hostname = stub["fqdn"],
@@ -102,25 +124,37 @@ class NuoDBCluster:
       template = string.Template(f.read())
       f.close()
       userdata = template.substitute(template_vars)
-      obj = stub['obj'].create(ami=stub['ami'], key_name=self.ssh_key, instance_type=self.instance_type, security_group_ids=stub['security_group_ids'], subnet = stub['subnet'], getPublicAddress = True, userdata = userdata)
-      port = obj.agentPort
-      print "Waiting for %s to start" % obj.ext_fqdn
+      obj = stub['obj'].create(ami=stub['ami'], instance_type=instance_type, security_group_ids=stub['security_group_ids'], subnet = stub['subnet'], getPublicAddress = True, userdata = userdata)
+      print ("Waiting for %s to start" % obj.ext_fqdn),
       if obj.status() != "running":
-        sys.stdout.write(".")
+        print("."),
         time.sleep(30) #Wait 30 seconds in between node starts
       print
       obj.update_data()
-      print "Setting DNS for %s " % obj.ext_fqdn
-      obj.dns_set()
-      print "Waiting for agent on %s " % obj.ext_fqdn
-      while not obj.agent_running():
-        sys.stdout.write(".")
-        time.sleep(10)
-      print
-      self.sync()
+      if not self.dns_emulate:
+        print "Setting DNS for %s " % obj.ext_fqdn
+        obj.dns_set()
+      if wait_for_health:
+        healthy = False
+        count = 0
+        tries = 60
+        wait = 10
+        print "Waiting for agent on %s " % obj.ext_fqdn
+        while not healthy or count == tries:
+          if obj.agent_running():
+            healthy = True
+          else:
+            print("."),
+            time.sleep(wait)
+          count += 1
+        if not healthy:
+          print "Cannot reach agent on %s after %s seconds. Check firewalls and the host for errors." % (obj.ext_fqdn, str(tries * wait))
+          exit(1)
+        print
+      else:
+        print "Not waiting for agent on %s, node will come up asynchronously." % obj.ext_fqdn
       return obj
-        
-        
+             
     def connect_zone(self, zone):
       self.zones[zone] = nuodbaws.NuoDBzone(zone)
       self.zones[zone].connect(aws_access_key=self.aws_access_key, aws_secret=self.aws_secret)
@@ -135,41 +169,71 @@ class NuoDBCluster:
       for host in self.get_hosts():
         obj = self.get_host(host)
         zone = obj.region
+        wait_for_health = False
         if obj.isBroker == True:
+          # If this node is a broker, then pair it with brokers outside its region if you can
+          wait_for_health = True
           brokers = []
           for idx, azone in enumerate(self.get_zones()):
             if azone != zone:
-              brokers = self.db['customers'][self.cluster_name]['zones'][azone]['brokers']
+              for broker in self.db['customers'][self.cluster_name]['zones'][azone]['brokers']:
+                brokers.append(broker)
           if len(brokers) == 0:
           # There are no other brokers in other regions found. Add another peer in this region if there is one
               brokers = self.db['customers'][self.cluster_name]['zones'][zone]['brokers']
         else:
+          #If this node isn't a broker pair it with local zone brokers
           brokers = self.db['customers'][self.cluster_name]['zones'][zone]['brokers']
-        print host + " Setting peers to " + str(brokers)
+        print "%s: Setting peers to [%s]" % (host, ",".join(brokers))
         self.db['customers'][self.cluster_name]['zones'][zone]['hosts'][host]['chef_data']['nuodb']['brokers'] = brokers
-        self.sync()
-        self.__boot_host(host, zone)
- 
+        self.__boot_host(host, zone, wait_for_health = wait_for_health)
+      if self.dns_emulate:
+        self.set_dns_emulation()
+
+      
     def delete_db(self):
       self.exit()
-      os.remove(self.database_file)
+      if os.path.exists(self.database_file):
+        os.remove(self.database_file)
+    
+    def delete_dns(self, zone = None):
+      if zone == None:
+        zones = self.get_zones()
+      else:
+        zones = [zone]
+      for zone in zones:
+        hosts = self.get_hosts(zone=zone)
+        for host in hosts:
+          host_obj = self.get_host(host)
+          host_obj.dns_delete()
       
     def dump_db(self):
       return self.db
     
-    def exit(self):
-      self.db.close()
-    
     def get_brokers(self):
       try:
-        return self.db['customers'][self.cluster_name]['brokers']
+        brokers = []
+        for zone in self.get_zones():
+          for broker in self.db['customers'][self.cluster_name]['zones'][zone]['brokers']:
+            brokers.append(broker)
+        return brokers
       except:
         return []
-      
+    
     def get_host(self, host_id):
       name, customer, zone = host_id.split(".")
       if host_id in self.db['customers'][customer]['zones'][zone]['hosts']:
         return self.db['customers'][customer]['zones'][zone]['hosts'][host_id]['obj']
+      else:
+        raise Error("No host found with id of '%s'" % host_id)
+    
+    def get_host_address(self, host_id):
+      name, customer, zone = host_id.split(".")
+      if host_id in self.db['customers'][customer]['zones'][zone]['hosts']:
+        if self.dns_emulate:
+          return self.db['customers'][customer]['zones'][zone]['hosts'][host_id]['obj'].ext_ip
+        else:
+          return self.db['customers'][customer]['zones'][zone]['hosts'][host_id]['obj'].ext_fqdn
       else:
         raise Error("No host found with id of '%s'" % host_id)
     
@@ -190,8 +254,34 @@ class NuoDBCluster:
         zones.append(zone)
       return sorted(zones)
     
-    def sync(self):
-      self.db.sync()
+    def set_dns_emulation(self):
+      host_list = []
+      for host_id in self.get_hosts():
+        host = self.get_host(host_id)
+        host.update_data()
+        print("Waiting for an IP for %s" % host.ext_fqdn),
+        while len(host.ext_ip) == 0:
+          print ("."),
+          time.sleep(5)
+          host.update_data()
+        print("got %s" % host.ext_ip)
+        host_list.append([host.ext_fqdn, host.ext_ip])
+      for host_id in self.get_hosts():
+        host = self.get_host(host_id)
+        print ("Waiting for ssh on %s." % host.ext_fqdn),
+        while not host.is_port_available(22):
+          print ("."),
+          time.sleep(5)
+        print
+        for line in host_list:
+          hostname = line[0]
+          ip = line[1]
+          command = "sudo awk -v s=\"%s    %s\" '/%s/{f=1;$0=s}7;END{if(!f)print s}' /etc/hosts > /tmp/hosts && sudo chown root:root /tmp/hosts && sudo chmod 644 /tmp/hosts && sudo mv /tmp/hosts /etc/hosts" % (ip, hostname, hostname)
+          (rc, stdout, stderr) = host.ssh_execute(command)
+          if rc != 0:
+            print "Unable to set DNS emulation for %s: %s" % (host.fqdn, stderr)
+        host.agent_action(action = "restart")
+        host.webconsole_action(action = "restart")
       
     def terminate_hosts(self, zone = None):
       if zone == None:
@@ -202,12 +292,13 @@ class NuoDBCluster:
         hosts = self.get_hosts(zone=zone)
         for host in hosts:
           host_obj = self.get_host(host)
-          host_obj.terminate()
-          del self.db['customers'][self.cluster_name]['zones'][zone]['hosts'][host_obj.name]
-          for idx, broker in enumerate(self.db['customers'][self.cluster_name]['brokers']):
-            if zone in broker:
-              del self.db['customers'][self.cluster_name]['brokers'][idx]
-        self.db['customers'][self.cluster_name]['zones'][zone]['brokers'] = []
+          if host_obj.exists:
+            print "Terminating %s" % host
+            host_obj.terminate()
+            del self.db['customers'][self.cluster_name]['zones'][zone]['hosts'][host_obj.name]
+            for idx, broker in enumerate(self.db['customers'][self.cluster_name]['brokers']):
+              if zone in broker:
+                del self.db['customers'][self.cluster_name]['brokers'][idx]
      
 class Error(Exception):
   pass
